split InjectionContainer definition for SDK and app
